feat(upload): move encryptor + cache-seeder into homebase-upload; decouple from chat (#844 PR2)
- Move PayloadBundleEncryptor, PayloadBundleEncryptionService, PayloadCacheSeeder
  into id.homebase.upload (via git mv, history preserved).
- Decouple the encryptor from the chat module:
  - new UploadProtocol.DEFAULT_PAYLOAD_DESCRIPTOR_KEY (single source of truth);
    ChatProtocol delegates to it so the 6 chat call sites are unchanged.
  - new VideoEncodePolicy interface replaces the encryptor's UserPreferences dep.
    Wired in AppModule via a small adapter over UserPreferences, so homebase-common
    stays decoupled from homebase-upload.
- homebase-upload build: add kotlinx-coroutines-core + kermit (implementation, since
  homebase-api exposes them non-transitively) and the ExperimentalUuidApi opt-in.
- Update all importers across chat + core; rebind DI.
- Simplify the fail-soft test to an inline VideoEncodePolicy (drops the InMemorySettings
  fake and the now-unused multiplatform-settings test dep from homebase-chat).

MessageAttachmentBuilder/MessageThumbnailGenerator deliberately stay in chat — they
consume chat's AttachmentInput and are feature-side source prep; they already import
PayloadBundle from the module.

Verified: JVM/Android/wasmJs compile clean (module + chat + core + tests); fail-soft,
cache-seeder, sender tests pass. iOS unverified locally (CI on #908 covers it).
