+++ homebase-chat/build.gradle.kts
@@ -140,7 +140,6 @@ kotlin {
             implementation(libs.kotlinx.coroutines.test)
             implementation(libs.ktor.client.mock)
             implementation(libs.okio.fakefilesystem)
-            implementation(libs.multiplatform.settings)
         }
     }
