+++ homebase-chat/src/jvmTest/kotlin/id/homebase/chat/services/PayloadCacheSeederTest.kt
@@ -1,4 +1,5 @@
 package id.homebase.chat.services
+import id.homebase.upload.PayloadCacheSeeder
 import id.homebase.upload.PayloadBundle
 
 import id.homebase.api.client.auth.ApiCredentials
