+++ homebase-chat/src/jvmTest/kotlin/id/homebase/chat/services/convo/ConversationServiceTestFixture.kt
@@ -21,7 +21,7 @@ import id.homebase.api.sync.database.OutboxSync
 import id.homebase.api.sync.database.OutboxUploader
 import id.homebase.chat.services.ChatProtocol
 import id.homebase.upload.PayloadBundle
-import id.homebase.chat.services.PayloadBundleEncryptor
+import id.homebase.upload.PayloadBundleEncryptor
 import id.homebase.chat.services.SendMessageResult
 import id.homebase.chat.services.StatusMessageData
 import id.homebase.chat.services.outbox.OptimisticWriter
